requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-celery-beat==2.4.0
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.4
gevent==22.10.2
greenlet==2.0.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycparser==2.21
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
redis==4.4.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
zope.event==4.6
zope.interface==5.5.2
